Add files via upload
Performing data cleaning and exploratory data analysis (EDA) involves preparing and examining a dataset, such as the Titanic dataset from Kaggle or a car dataset. This process aims to understand the data, identify missing values, outliers, and inconsistencies, and explore relationships between variables. Through EDA, patterns, trends, and insights in the data are discovered, aiding in decision-making and further analysis.